+++ radio/src/gui/128x64/model_templates.cpp
@@ -40,7 +40,7 @@ void menuModelTemplates(event_t event)
 
   coord_t y = MENU_HEADER_HEIGHT + 1;
   uint8_t k = 0;
-  for (uint8_t i=0; i<LCD_LINES-1 && k<TMPL_COUNT; i++) {
+  for (uint32_t i=0; i<LCD_LINES-1 && k<TMPL_COUNT; i++) {
     k = i+menuVerticalOffset;
     lcdDrawNumber(3*FW, y, k, (sub==k ? INVERS : 0)|LEADING0, 2);
     lcdDrawTextAtIndex(4*FW, y, STR_VTEMPLATES, k, (sub==k ? INVERS  : 0));
