+++ radio/src/gui/212x64/model_flightmodes.cpp
@@ -61,7 +61,7 @@ void menuModelFlightModesAll(event_t event)
     drawColumnHeader(STR_PHASES_HEADERS, posHorz);
   }
 
-  for (uint8_t i=0; i<LCD_LINES-1; i++) {
+  for (uint32_t i=0; i<LCD_LINES-1; i++) {
     coord_t y = MENU_HEADER_HEIGHT + 1 + i*FH;
     uint8_t k = i+menuVerticalOffset;
